requirements.txt
praw>=7.7.1
pandas>=2.2.0
numpy>=1.26.0
nltk>=3.8.1
spacy>=3.7.2
gensim>=4.3.2
vaderSentiment>=3.3.2
matplotlib>=3.8.0
seaborn>=0.13.0
wordcloud>=1.9.2
pyLDAvis>=3.4.1
scikit-learn>=1.4.0
pyyaml>=6.0.1
tqdm>=4.66.1
jupyter>=1.0.0
python-dotenv>=1.0.0
openpyxl>=3.1.2
plotly>=5.17.0
